backend/requirements.txt
fastapi==0.112.2
psycopg2==2.9.9
pydantic==2.8.2
PyJWT==2.9.0
PyJWT==2.9.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
